pyproject.toml
[project]
name = "openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "altair>=5.5.0",
  "duckdb>=1.3.2",
  "faker>=37.4.2",
  "httpx>=0.28.1",
  "marimo>=0.14.12",
  "nbformat>=5.10.4",
  "openai-agents[viz]>=0.2.2",
  "pandas>=2.3.1",
  "pandera>=0.25.0",
  "polars[pyarrow]>=1.31.0",
  "pre-commit>=4.2.0",
  "pydantic>=2.11.7",
  "pypdf>=5.8.0",
  "pytest>=8.4.1",
  "python-lsp-ruff>=2.2.2",
  "python-lsp-server>=1.13.0",
  "sqlglot>=27.1.0",
  "tqdm>=4.67.1",
  "vegafusion>=2.0.2",
  "vl-convert-python>=1.8.0",
  "websockets>=15.0.1",
]

[dependency-groups]
dev = ["ruff>=0.12.4"]

[tool.ruff]
fix = true
unsafe-fixes = false
exclude = ["notebooks/"]
